+++ src/main/token-service/TokenService.ts
@@ -1,7 +1,7 @@
 import { GetAccessTokenRequest, GrantType, RenewTokenRequest, TokenClaim, TokenHeader, TokenIntrospectionRequest } from "./TokenService.model";
 import { Helper, CustomException } from "../common/Helper";
 import { JwtHelper } from "../common/JwtHelper";
-import { LoginPrecheckRequest } from "../common/Common.model";
+import { HTTPRequestHeader, LoginPrecheckRequest } from "../common/Common.model";
 
 /**
  * To get a new token with the grant type refresh_token, call **renewToken()**.
@@ -110,9 +110,9 @@ export function validateAccessToken(options: TokenIntrospectionRequest) {
  * });
  * ```
  */
-export function loginPrecheck(options: LoginPrecheckRequest) {
+export function loginPrecheck(options: LoginPrecheckRequest, headers?: HTTPRequestHeader) {
   const _serviceURL = window.webAuthSettings.authority + "/token-srv/prelogin/metadata/" + options.track_id;
-  return Helper.createHttpPromise(undefined, _serviceURL, false, "GET");
+  return Helper.createHttpPromise(undefined, _serviceURL, false, "GET", undefined, headers);
 }
 
 /**
@@ -130,9 +130,9 @@ export function loginPrecheck(options: LoginPrecheckRequest) {
  * });
  * ```
  */
-export function getMissingFields(trackId: string) {
+export function getMissingFields(trackId: string, headers?: HTTPRequestHeader) {
   const _serviceURL = window.webAuthSettings.authority + "/token-srv/prelogin/metadata/" + trackId;
-  return Helper.createHttpPromise(undefined, _serviceURL, false, "GET");
+  return Helper.createHttpPromise(undefined, _serviceURL, false, "GET", undefined, headers);
 }
 
 /**