+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## V4.3.1
+
+### Changed
+- Update required functions to accept headers
+- expand resetPassword flow functions with optional handleResponseAsJson for compatibility with older cidaas version
+
 ## V4.3.0
 
 ### Added